direct-requirements.txt
## NOTES: Do not overwrite this file by `pip freeze`
## Just add packages if they are used directly in codes
## This file does not include sub dependencies

## Libs
libclang==14.0.6
libmagic==1.0
libsvm==3.23.0.4
lxml==4.9.2
matplotlib==3.4.2
aiofiles>=22.1.0

## Packages
cachetools>=4.2.2
dynaconf>=3.1.9
email-validator>=1.2.1
fastapi>=0.82.0
minio>=7.1.11
xmltodict>=0.13.0
tabula-py>=2.5.1
pytesseract==0.3.7
tesseract>=0.1.3
tifffile>=2021.6.14
typing-extensions>=3.10
opencv-python==4.5.2.54
gunicorn==20.1.0
uuid==1.30
uvicorn==0.14.0
numpy==1.20.3
redis==4.4.2
scikit-image>=0.19.3
image-quality>=1.2.7
deskew==1.3.2
## install python-magic-bin by github
python-magic @ git+https://github.com/julian-r/python-magic@b84a85a123ca9896df4a5df812a9b4f7f93335cc
python-docx==0.8.11
python-multipart==0.0.5
pikepdf==5.5.0
reportlab==3.6.11
Pillow==9.2.0
requests>=2.28.1
urllib3>=1.26.12
bs4>=0.0.1
pdfminer.six==20221105
fastapi-limiter==0.1.5
async-timeout==4.0.2

## Google
firebase-admin>=5.0.1
google-cloud-vision>=3.1.1
google-cloud-firestore>=2.6.1
google-cloud-storage>=2.5.0
google-cloud-translate>=3
googletrans==3.1.0a0
google-auth==1.32.0
google-auth-httplib2==0.1.0

## In use or dev tools
OPA-python-client==1.3.3
pip-autoremove==0.10.0
pyparsing==2.4.7

# Redis queue
rq==1.11.1

# url-malicious
# These will be installed through CONDA
# wordcloud==1.8.2.2
scikit-learn==1.2.2
# joblib==1.1.0
# pandas==1.4.2
# matplotlib==3.5.3
# seaborn==0.12.2
# pip install wordcloud scikit-learn joblib pandas matplotlib seaborn trafilatura

# main text 
trafilatura==1.4.1

# Vertex AI
google-cloud-aiplatform==1.23.0